refactor: 검색 결과 패키지 경로 변경
