+++ android/app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.Turip"
         tools:targetApi="31">
         <activity
-            android:name=".ui.searchResult.SearchResultActivity"
+            android:name=".ui.search.result.SearchResultActivity"
             android:exported="false" />
         <activity
             android:name=".ui.main.MainActivity"
