+++ android/app/src/main/res/layout/activity_search_result.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.search.SearchResultActivity">
+    tools:context=".ui.search.result.SearchResultActivity">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/tb_search_result"
